refactor(KeyboardShortcuts): move code into separate folder
